# Handle bool defaults only when field missing