## Summary
- adjust `setDefaults` to check if form field is provided before applying default values
- add `fieldProvided` helper
- ensure decoder uses new behaviour and update recursive calls
- test bool default behaviour with missing vs provided values

## Testing
- `go test ./...`