@ReneWerner87 2025-05-25
```
                                     │   old.txt    │                new.txt                │
                                     │    sec/op    │    sec/op      vs base                │
All-12                                 77.14n ± ∞ ¹    78.35n ± ∞ ¹       ~ (p=0.700 n=3) ²
DecoderMultipartFiles-12               5.078µ ± ∞ ¹    5.101µ ± ∞ ¹       ~ (p=1.000 n=3) ²
IsMultipartFile/IsMultipartFile-0-12   3.625n ± ∞ ¹    3.595n ± ∞ ¹       ~ (p=0.400 n=3) ²
IsMultipartFile/IsMultipartFile-1-12   3.629n ± ∞ ¹    3.607n ± ∞ ¹       ~ (p=1.000 n=3) ²
IsMultipartFile/IsMultipartFile-2-12   3.025n ± ∞ ¹    2.992n ± ∞ ¹       ~ (p=0.400 n=3) ²
IsMultipartFile/IsMultipartFile-3-12   3.638n ± ∞ ¹    3.680n ± ∞ ¹       ~ (p=0.700 n=3) ²
IsMultipartFile/IsMultipartFile-4-12   2.126n ± ∞ ¹    2.141n ± ∞ ¹       ~ (p=1.000 n=3) ²
IsMultipartFile/IsMultipartFile-5-12   3.618n ± ∞ ¹    3.786n ± ∞ ¹       ~ (p=0.700 n=3) ²
IsMultipartFile/IsMultipartFile-6-12   6.647n ± ∞ ¹    6.770n ± ∞ ¹       ~ (p=0.400 n=3) ²
HandleMultipartField-12                98.23n ± ∞ ¹   102.10n ± ∞ ¹       ~ (p=0.200 n=3) ²
LargeStructDecode-12                   5.785µ ± ∞ ¹    5.813µ ± ∞ ¹       ~ (p=0.400 n=3) ²
LargeStructDecodeParallel-12           2.786µ ± ∞ ¹    2.911µ ± ∞ ¹       ~ (p=0.100 n=3) ²
SimpleStructDecode-12                  2.470µ ± ∞ ¹    2.611µ ± ∞ ¹       ~ (p=0.100 n=3) ²
CheckRequiredFields-12                 868.6n ± ∞ ¹    847.7n ± ∞ ¹       ~ (p=0.400 n=3) ²
TimeDurationDecoding-12                519.8n ± ∞ ¹    522.1n ± ∞ ¹       ~ (p=0.400 n=3) ²
SimpleStructEncode-12                  1.306µ ± ∞ ¹    1.280µ ± ∞ ¹       ~ (p=0.600 n=3) ²
SimpleStructEncodeParallel-12          193.3n ± ∞ ¹    192.8n ± ∞ ¹       ~ (p=1.000 n=3) ²
LargeStructEncode-12                   2.293µ ± ∞ ¹    2.265µ ± ∞ ¹       ~ (p=0.400 n=3) ²
LargeStructEncodeParallel-12           369.8n ± ∞ ¹    345.4n ± ∞ ¹       ~ (p=0.100 n=3) ²
TimeDurationEncoding-12                230.4n ± ∞ ¹    228.7n ± ∞ ¹       ~ (p=1.000 n=3) ²

```
benchmark is not really changed